reduce memory usage

+++ src/graphics/dash_anim.h
@@ -2,17 +2,17 @@
 #define DASH_ANIM_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 struct DashFrame
 {
     int _id;
-    // This is the rigging for the frame.
     int framemap_id;
     int translator_count;
-    int* index_frame_ids;
-    int* translator_arg_x;
-    int* translator_arg_y;
-    int* translator_arg_z;
+    int16_t* index_frame_ids;
+    int16_t* translator_arg_x;
+    int16_t* translator_arg_y;
+    int16_t* translator_arg_z;
     bool showing;
 };
